+++ mlops/common/pipeline_utils.py
@@ -1,0 +1,172 @@
+"""
+This module defines a machine learning pipeline for processing, training, and evaluating data.
+"""
+from azure.identity import DefaultAzureCredential
+from azure.ai.ml.dsl import pipeline
+from azure.ai.ml import MLClient
+import time
+
+from mlops.common.config_utils import MLOpsConfig
+from mlops.common.get_compute import get_compute
+from mlops.common.naming_utils import generate_experiment_name, generate_run_name
+
+
+def set_pipeline_properties(
+    pipeline_job: pipeline,
+    cluster_name: str,
+    display_name: str,
+    build_environment: str,
+    build_reference: str,
+):
+    """
+    """
+
+    pipeline_job.display_name = display_name
+    pipeline_job.tags = {
+        "environment": build_environment,
+        "build_reference": build_reference,
+    }
+
+    # set pipeline level compute
+    pipeline_job.settings.default_compute = cluster_name
+    pipeline_job.settings.force_rerun = True
+    # set pipeline level datastore
+    pipeline_job.settings.default_datastore = "workspaceblobstore"
+
+    return pipeline_job
+
+def execute_pipeline(
+    subscription_id: str,
+    resource_group_name: str,
+    workspace_name: str,
+    experiment_name: str,
+    pipeline_job: pipeline,
+    wait_for_completion: str,
+    output_file: str,
+):
+    """
+    Execute a pipeline job in Azure Machine Learning service.
+
+    Args:
+        subscription_id (str): The Azure subscription ID.
+        resource_group_name (str): The name of the resource group.
+        workspace_name (str): The name of the Azure Machine Learning workspace.
+        experiment_name (str): The name of the experiment.
+        pipeline_job (pipeline): The pipeline job to be executed.
+        wait_for_completion (str): "True" or "False" indicates whether to wait for the job to complete.
+        output_file (str): The path to the output file where the job name will be written.
+
+    Raises:
+        Exception: If the job fails to complete.
+
+    Returns:
+        None
+    """
+    try:
+        client = MLClient(
+            DefaultAzureCredential(),
+            subscription_id=subscription_id,
+            resource_group_name=resource_group_name,
+            workspace_name=workspace_name,
+        )
+
+        pipeline_job = client.jobs.create_or_update(
+            pipeline_job, experiment_name=experiment_name
+        )
+
+        print(f"The job {pipeline_job.name} has been submitted!")
+        if output_file is not None:
+            with open(output_file, "w") as out_file:
+                out_file.write(pipeline_job.name)
+
+        if wait_for_completion == "True":
+            total_wait_time = 3600
+            current_wait_time = 0
+            job_status = [
+                "NotStarted",
+                "Queued",
+                "Starting",
+                "Preparing",
+                "Running",
+                "Finalizing",
+                "Provisioning",
+                "CancelRequested",
+                "Failed",
+                "Canceled",
+                "NotResponding",
+            ]
+
+            while pipeline_job.status in job_status:
+                if current_wait_time <= total_wait_time:
+                    time.sleep(20)
+                    pipeline_job = client.jobs.get(pipeline_job.name)
+
+                    print("Job Status:", pipeline_job.status)
+
+                    current_wait_time = current_wait_time + 15
+
+                    if (
+                        pipeline_job.status == "Failed"
+                        or pipeline_job.status == "NotResponding"
+                        or pipeline_job.status == "CancelRequested"
+                        or pipeline_job.status == "Canceled"
+                    ):
+                        break
+                else:
+                    break
+
+            if pipeline_job.status == "Completed" or pipeline_job.status == "Finished":
+                print("job completed")
+            else:
+                raise Exception("Sorry, exiting job with failure..")
+    except Exception as ex:
+        print(
+            "Oops! invalid credentials or error while creating ML environment.. Try again...",
+            ex,
+        )
+        raise
+
+
+def prepare_and_execute_pipeline(pipeline, model_name, build_environment, wait_for_completion, output_file):
+
+    config = MLOpsConfig(environment=build_environment)
+
+    ml_client = MLClient(
+        DefaultAzureCredential(),
+        config.aml_config["subscription_id"],
+        config.aml_config["resource_group_name"],
+        config.aml_config["workspace_name"],
+    )
+
+    pipeline_config = config.get_pipeline_config(model_name)
+    published_experiment_name = generate_experiment_name(model_name)
+    published_run_name = generate_run_name(config.environment_configuration["build_reference"])
+
+    compute = get_compute(
+        config.aml_config["subscription_id"],
+        config.aml_config["resource_group_name"],
+        config.aml_config["workspace_name"],
+        pipeline_config["cluster_name"],
+        pipeline_config["cluster_size"],
+        pipeline_config["cluster_region"],
+    )
+
+    pipeline_job = pipeline.construct_pipeline(ml_client)
+
+    pipeline_job = set_pipeline_properties(
+        pipeline_job,
+        cluster_name=compute.name,
+        display_name=published_run_name,
+        build_environment=build_environment,
+        build_reference=model_name,
+    )
+
+    execute_pipeline(
+        config.aml_config["subscription_id"],
+        config.aml_config["resource_group_name"],
+        config.aml_config["workspace_name"],
+        published_experiment_name,
+        pipeline_job,
+        wait_for_completion,
+        output_file,
+    )
