+++ mlops/london_taxi/src/mlops_pipeline.py
@@ -10,47 +10,28 @@ The pipeline executes the following steps in order:
 6. Finalize and Persist Model: Handles tasks like persisting model metadata, registering the model,
 and generating reports.
 """
-from azure.identity import DefaultAzureCredential
+
 import argparse
 from azure.ai.ml.dsl import pipeline
-from azure.ai.ml import MLClient, Input
+from azure.ai.ml import Input
 from azure.ai.ml import load_component
-import time
 import os
-from mlops.common.get_compute import get_compute
 from mlops.common.get_environment import get_environment
 from mlops.common.config_utils import MLOpsConfig
-from mlops.common.naming_utils import (
-    generate_experiment_name,
-    generate_model_name,
-    generate_run_name,
-)
-
+from mlops.common.naming_utils import generate_model_name
+from mlops.common.pipeline_utils import prepare_and_execute_pipeline
 
 gl_pipeline_components = []
 
 
 @pipeline()
 def london_taxi_data_regression(pipeline_job_input, model_name, build_reference):
-    """
-    Run a pipeline for regression analysis on NYC taxi data.
-
-    Parameters:
-    pipeline_job_input (str): Path to the input data.
-    model_name (str): Name of the model.
-    build_reference (str): Reference for the build.
-
-    Returns:
-    dict: A dictionary containing paths to various data, the model, predictions, and score report.
-    """
-    prepare_sample_data = gl_pipeline_components[0](
-        raw_data=pipeline_job_input,
-    )
+    prepare_sample_data = gl_pipeline_components[0](raw_data=pipeline_job_input)
     transform_sample_data = gl_pipeline_components[1](
-        clean_data=prepare_sample_data.outputs.prep_data,
+        clean_data=prepare_sample_data.outputs.prep_data
     )
     train_with_sample_data = gl_pipeline_components[2](
-        training_data=transform_sample_data.outputs.transformed_data,
+        training_data=transform_sample_data.outputs.transformed_data
     )
     predict_with_sample_data = gl_pipeline_components[3](
         model_input=train_with_sample_data.outputs.model_output,
@@ -77,208 +58,83 @@ def london_taxi_data_regression(pipeline_job_input, model_name, build_reference)
     }
 
 
-def construct_pipeline(
-    cluster_name: str,
-    environment_name: str,
-    display_name: str,
-    build_environment: str,
-    build_reference: str,
-    model_name: str,
-    dataset_name: str,
-    ml_client,
-):
+class London_Taxi:
     """
-    Construct a pipeline job for NYC taxi data regression.
-
-    Args:
-        cluster_name (str): The name of the cluster to use for pipeline execution.
-        environment_name (str): The name of the environment to use for pipeline execution.
-        display_name (str): The display name of the pipeline job.
-        build_environment (str): The environment to deploy the pipeline job.
-        build_reference (str): The build reference for the pipeline job.
-        model_name (str): The name of the model.
-        dataset_name (str): The name of the dataset.
-        ml_client: The machine learning client.
-
-    Returns:
-        pipeline_job: The constructed pipeline job.
+    This class defines the machine learning pipeline for processing, training, and evaluating data.
     """
-    registered_data_asset = ml_client.data.get(name=dataset_name, label="latest")
 
-    parent_dir = os.path.join(os.getcwd(), "mlops/london_taxi/components")
+    def __init__(self, environment_name, build_reference, model_name, dataset_name):
+        """
+        Initialize the pipeline job components.
 
-    prepare_data = load_component(source=parent_dir + "/prep.yml")
-    transform_data = load_component(source=parent_dir + "/transform.yml")
-    train_model = load_component(source=parent_dir + "/train.yml")
-    predict_result = load_component(source=parent_dir + "/predict.yml")
-    score_data = load_component(source=parent_dir + "/score.yml")
-    register_model = load_component(source=parent_dir + "/register.yml")
+        Args:
+            ml_client: The machine learning client.
+            environment_name (str): The name of the environment to use for pipeline execution.
+            build_reference (str): The build reference for the pipeline job.
+            model_name (str): The name of the model.
+            dataset_name (str): The name of the dataset.
+        """
+        self.environment_name = environment_name
+        self.build_reference = build_reference
+        self.model_name = model_name
+        self.dataset_name = dataset_name
+        self.gl_pipeline_components = []
 
-    # Set the environment name to custom environment using name and version number
-    prepare_data.environment = environment_name
-    transform_data.environment = environment_name
-    train_model.environment = environment_name
-    predict_result.environment = environment_name
-    score_data.environment = environment_name
-    register_model.environment = environment_name
+    def construct_pipeline(self, ml_client):
+        """
+        Construct a pipeline job for NYC taxi data regression.
 
-    gl_pipeline_components.append(prepare_data)
-    gl_pipeline_components.append(transform_data)
-    gl_pipeline_components.append(train_model)
-    gl_pipeline_components.append(predict_result)
-    gl_pipeline_components.append(score_data)
-    gl_pipeline_components.append(register_model)
-
-    pipeline_job = london_taxi_data_regression(
-        Input(type="uri_folder", path=registered_data_asset.id),
-        model_name,
-        build_reference,
-    )
-
-    pipeline_job.display_name = display_name
-    pipeline_job.tags = {
-        "environment": build_environment,
-        "build_reference": build_reference,
-    }
-
-    # demo how to change pipeline output settings
-    pipeline_job.outputs.pipeline_job_prepped_data.mode = "rw_mount"
-
-    # set pipeline level compute
-    pipeline_job.settings.default_compute = cluster_name
-    pipeline_job.settings.force_rerun = True
-    # set pipeline level datastore
-    pipeline_job.settings.default_datastore = "workspaceblobstore"
-
-    return pipeline_job
-
-
-def execute_pipeline(
-    subscription_id: str,
-    resource_group_name: str,
-    workspace_name: str,
-    experiment_name: str,
-    pipeline_job: pipeline,
-    wait_for_completion: str,
-    output_file: str,
-):
-    """
-    Execute a pipeline job in Azure Machine Learning service.
-
-    Args:
-        subscription_id (str): The Azure subscription ID.
-        resource_group_name (str): The name of the resource group.
-        workspace_name (str): The name of the Azure Machine Learning workspace.
-        experiment_name (str): The name of the experiment.
-        pipeline_job (pipeline): The pipeline job to be executed.
-        wait_for_completion (str): "True" or "False" - indicates whether to wait for the job to complete.
-        output_file (str): The path to the output file where the job name will be written.
-
-    Raises:
-        Exception: If the job fails to complete.
-
-    Returns:
-        None
-    """
-    try:
-        client = MLClient(
-            DefaultAzureCredential(),
-            subscription_id=subscription_id,
-            resource_group_name=resource_group_name,
-            workspace_name=workspace_name,
+        Returns:
+            pipeline_job: The constructed pipeline job components.
+        """
+        registered_data_asset = ml_client.data.get(
+            name=self.dataset_name, label="latest"
         )
 
-        pipeline_job = client.jobs.create_or_update(
-            pipeline_job, experiment_name=experiment_name
+        parent_dir = os.path.join(os.getcwd(), "mlops/london_taxi/components")
+
+        prepare_data = load_component(source=parent_dir + "/prep.yml")
+        transform_data = load_component(source=parent_dir + "/transform.yml")
+        train_model = load_component(source=parent_dir + "/train.yml")
+        predict_result = load_component(source=parent_dir + "/predict.yml")
+        score_data = load_component(source=parent_dir + "/score.yml")
+        register_model = load_component(source=parent_dir + "/register.yml")
+
+        # Set the environment name to custom environment using name and version number
+        prepare_data.environment = self.environment_name
+        transform_data.environment = self.environment_name
+        train_model.environment = self.environment_name
+        predict_result.environment = self.environment_name
+        score_data.environment = self.environment_name
+        register_model.environment = self.environment_name
+
+        gl_pipeline_components.append(prepare_data)
+        gl_pipeline_components.append(transform_data)
+        gl_pipeline_components.append(train_model)
+        gl_pipeline_components.append(predict_result)
+        gl_pipeline_components.append(score_data)
+        gl_pipeline_components.append(register_model)
+
+        pipeline_job = london_taxi_data_regression(
+            Input(type="uri_folder", path=registered_data_asset.id),
+            self.model_name,
+            self.build_reference,
         )
 
-        print(f"The job {pipeline_job.name} has been submitted!")
-        if output_file is not None:
-            with open(output_file, "w") as out_file:
-                out_file.write(pipeline_job.name)
+        # demo how to change pipeline output settings
+        pipeline_job.outputs.pipeline_job_prepped_data.mode = "rw_mount"
 
-        if wait_for_completion == "True":
-            total_wait_time = 3600
-            current_wait_time = 0
-            job_status = [
-                "NotStarted",
-                "Queued",
-                "Starting",
-                "Preparing",
-                "Running",
-                "Finalizing",
-                "Provisioning",
-                "CancelRequested",
-                "Failed",
-                "Canceled",
-                "NotResponding",
-            ]
-
-            while pipeline_job.status in job_status:
-                if current_wait_time <= total_wait_time:
-                    time.sleep(20)
-                    pipeline_job = client.jobs.get(pipeline_job.name)
-
-                    print("Job Status:", pipeline_job.status)
-
-                    current_wait_time = current_wait_time + 15
-
-                    if (
-                        pipeline_job.status == "Failed"
-                        or pipeline_job.status == "NotResponding"
-                        or pipeline_job.status == "CancelRequested"
-                        or pipeline_job.status == "Canceled"
-                    ):
-                        break
-                else:
-                    break
-
-            if pipeline_job.status == "Completed" or pipeline_job.status == "Finished":
-                print("job completed")
-            else:
-                raise Exception("Sorry, exiting job with failure..")
-    except Exception as ex:
-        print(
-            "Oops! invalid credentials or error while creating ML environment.. Try again...",
-            ex,
-        )
-        raise
+        return pipeline_job
 
 
 def prepare_and_execute(
-    build_environment: str,
-    wait_for_completion: str,
-    output_file: str,
+    model_name, build_environment, wait_for_completion, output_file
 ):
-    """
-    Prepare and execute the MLOps pipeline.
-
-    Args:
-        build_environment (str): environment name to execute.
-        wait_for_completion (str): "True" or "False" - indicates whether to wait for the job to complete.
-        output_file (str): The path to the output file where the job name will be written.
-    """
-    model_name = "london_taxi"
 
     config = MLOpsConfig(environment=build_environment)
 
-    ml_client = MLClient(
-        DefaultAzureCredential(),
-        config.aml_config["subscription_id"],
-        config.aml_config["resource_group_name"],
-        config.aml_config["workspace_name"],
-    )
-
     pipeline_config = config.get_pipeline_config(model_name)
-
-    compute = get_compute(
-        config.aml_config["subscription_id"],
-        config.aml_config["resource_group_name"],
-        config.aml_config["workspace_name"],
-        pipeline_config["cluster_name"],
-        pipeline_config["cluster_size"],
-        pipeline_config["cluster_region"],
-    )
+    published_model_name = generate_model_name(model_name)
 
     environment = get_environment(
         config.aml_config["subscription_id"],
@@ -291,37 +147,24 @@ def prepare_and_execute(
 
     print(f"Environment: {environment.name}, version: {environment.version}")
 
-    published_model_name = generate_model_name(model_name)
-    published_experiment_name = generate_experiment_name(model_name)
-    published_run_name = generate_run_name(
-        config.environment_configuration["build_reference"]
-    )
-
-    pipeline_job = construct_pipeline(
-        compute.name,
+    pipeline = London_Taxi(
         f"azureml:{environment.name}:{environment.version}",
-        published_run_name,
-        build_environment,
         config.environment_configuration["build_reference"],
         published_model_name,
         pipeline_config["dataset_name"],
-        ml_client,
     )
 
-    execute_pipeline(
-        config.aml_config["subscription_id"],
-        config.aml_config["resource_group_name"],
-        config.aml_config["workspace_name"],
-        published_experiment_name,
-        pipeline_job,
-        wait_for_completion,
-        output_file,
+    prepare_and_execute_pipeline(
+        pipeline, model_name, build_environment, wait_for_completion, output_file
     )
 
 
 def main():
     """Parse the command line arguments and call the `prepare_and_execute` function."""
     parser = argparse.ArgumentParser("build_environment")
+    parser.add_argument(
+        "--model_name", type=str, help="name of the model", default="london_taxi"
+    )
     parser.add_argument(
         "--build_environment",
         type=str,
@@ -339,7 +182,10 @@ def main():
     args = parser.parse_args()
 
     prepare_and_execute(
-        args.build_environment, args.wait_for_completion, args.output_file
+        args.model_name,
+        args.build_environment,
+        args.wait_for_completion,
+        args.output_file,
     )
