+++ mlops/london_taxi/start_local_pipeline.py
@@ -2,4 +2,4 @@
 from mlops.london_taxi.src import mlops_pipeline
 
 if __name__ == "__main__":
-    mlops_pipeline.prepare_and_execute("pr", "True", None)
+    mlops_pipeline.prepare_and_execute("london_taxi", "pr", "True", None)
